[android] refactor turn res code
