+++ android/app/src/main/java/app/organicmaps/routing/NavigationController.java
@@ -20,10 +20,8 @@ import app.organicmaps.maplayer.MapButtonsViewModel;
 import app.organicmaps.sdk.Framework;
 import app.organicmaps.sdk.Router;
 import app.organicmaps.sdk.maplayer.traffic.TrafficManager;
-import app.organicmaps.sdk.routing.CarDirection;
 import app.organicmaps.sdk.routing.RoutingController;
 import app.organicmaps.sdk.routing.RoutingInfo;
-import app.organicmaps.sdk.util.RoundaboutExit;
 import app.organicmaps.sdk.util.StringUtils;
 import app.organicmaps.util.UiUtils;
 import app.organicmaps.util.Utils;
@@ -113,15 +111,13 @@ public class NavigationController implements TrafficManager.TrafficCallback, Nav
   private void updateVehicle(@NonNull RoutingInfo info)
   {
     mNextTurnDistance.setText(Utils.formatDistance(mFrame.getContext(), info.distToTurn));
+    mNextTurnImage.setImageResource(info.carDirection.getTurnRes(info.exitNum));
 
-    if (CarDirection.isRoundAbout(info.carDirection))
-      mNextTurnImage.setImageResource(RoundaboutExit.getRes(info.exitNum));
-    else
-      info.carDirection.setTurnDrawable(mNextTurnImage);
-
-    UiUtils.showIf(info.nextCarDirection.containsNextTurn(), mNextNextTurnFrame);
     if (info.nextCarDirection.containsNextTurn())
-      info.nextCarDirection.setNextTurnDrawable(mNextNextTurnImage);
+    {
+      mNextNextTurnImage.setImageResource(info.nextCarDirection.getNextTurnRes());
+      UiUtils.show(mNextNextTurnFrame);
+    }
 
     mLanesView.setLanes(info.lanes);
 
@@ -131,8 +127,7 @@ public class NavigationController implements TrafficManager.TrafficCallback, Nav
   private void updatePedestrian(@NonNull RoutingInfo info)
   {
     mNextTurnDistance.setText(Utils.formatDistance(mFrame.getContext(), info.distToTurn));
-
-    info.pedestrianTurnDirection.setTurnDrawable(mNextTurnImage);
+    mNextTurnImage.setImageResource(info.pedestrianTurnDirection.getTurnRes());
   }
 
   public void updateNorth()
